server/requirements.txt
Django==5.1.5
PyJWT==2.10.1
PyYAML==6.0.2
anyio==4.8.0
apify_client==1.8.1
apify_shared==1.2.1
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-yasg==1.21.8
exceptiongroup==1.2.2
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-auth==2.37.0
googleapis-common-protos==1.66.0
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
more-itertools==10.6.0
numpy==2.2.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-environ==0.4.54
python3-openid==3.2.0
pytz==2024.2
requests-oauthlib==2.0.0
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
